feat: list-of-groups refactor + doubleDate as component
