+++ src/components/double-date-filter/index.js
@@ -1,0 +1,1 @@
+export { DoubleDateFilter } from './doubleDateFilter.js'
